Rename credentials file to credential
